requirements.txt
## The following requirements were added by pip freeze:
Django==3.2.19
pytz==2023.3
djangorestframework==3.14.0
psycopg2==2.9.4
psycopg2-binary==2.9.5
urllib3==1.26.15
langchain==0.0.160
openai==0.27.6
gunicorn==20.1.0
dj-database-url==2.0.0
whitenoise==5.0.1
